Change architecture of Repository to match nrf5x-firmware repository.
The architecture of repository is changed to match with nrf5x repository so
that it'll be easier for developer to develop firmware on two different
chip-sets with same architecture.

This commit contains simple firmware for SenseBe Tx. In this firmware user
can switch between short, medium, long range for IR beam by button press.

Signed-off-by: Tejas Vasekar <tsvasekar@gmail.com>